# api/openapi.yaml
openapi: 3.0.3
info:
  title: ShopList API
  version: 0.1.0
  description: Контракт API. Описано то, что уже реализовано; остальное добавляется по мере реализации.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Проверка работоспособности (включая базу данных)
      responses:
        "200":
          description: Всё работает
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
        "503":
          description: База данных недоступна
  /api/v1/catalog/search:
    get:
      summary: Поиск товара в общем справочнике (с учётом опечаток)
      parameters:
        - { name: q, in: query, required: true, schema: { type: string }, example: молоко }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        "200":
          description: Найденные товары
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/CatalogItem" }
        "400":
          description: Не указан q или некорректный limit
  /api/v1/categories:
    get:
      summary: Категории товаров в порядке обхода магазина
      responses:
        "200":
          description: Список категорий
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Category" }
  /api/v1/auth/request-code:
    post:
      summary: Отправить код входа на email (аккаунт создаётся при первом входе)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "204": { description: Код отправлен }
        "400": { description: Некорректный email }
        "429": { description: Код запрашивали меньше минуты назад (см. Retry-After) }
  /api/v1/auth/verify:
    post:
      summary: Проверить код и получить токены
      description: Код действует 10 минут, после 5 неверных попыток блокируется.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email }
                code: { type: string, example: "123456" }
      responses:
        "200":
          description: Токены
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tokens" }
        "401": { description: Неверный или просроченный код }
  /api/v1/auth/refresh:
    post:
      summary: Обменять refresh-токен на новую пару токенов
      description: Refresh-токен одноразовый. Повторное использование старого отзывает все токены пользователя.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: Новая пара токенов
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tokens" }
        "401": { description: Токен недействителен }
  /api/v1/auth/logout:
    post:
      summary: Выйти (отозвать refresh-токен)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "204": { description: Готово }
  /api/v1/me:
    get:
      summary: Текущий пользователь
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Профиль
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { description: Нет или недействителен access-токен }
    patch:
      summary: Изменить имя
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 50 }
      responses:
        "200":
          description: Обновлённый профиль
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { description: Некорректное имя }
        "401": { description: Нет или недействителен access-токен }
    delete:
      summary: Удалить аккаунт
      description: >
        Требование App Store и Google Play. Общие списки переходят самому раннему из остальных участников,
        списки без других участников и остальные данные пользователя удаляются.
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Аккаунт удалён }
        "401": { description: Нет или недействителен access-токен }
  /api/v1/lists:
    get:
      summary: Списки, в которых состоит пользователь (свежие сверху)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Списки
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/List" }
    post:
      summary: Создать список (создатель становится владельцем)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, minLength: 1, maxLength: 100 }
      responses:
        "201":
          description: Созданный список
          content:
            application/json:
              schema: { $ref: "#/components/schemas/List" }
        "400": { description: Название пустое или длиннее 100 символов }
  /api/v1/lists/{listID}:
    parameters:
        - { name: listID, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Список. Для не участника всегда 404, чтобы не раскрывать существование списка
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Список
          content:
            application/json:
              schema: { $ref: "#/components/schemas/List" }
        "404": { description: Нет такого списка или вы в нём не участвуете }
    patch:
      summary: Переименовать (только владелец)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, minLength: 1, maxLength: 100 }
      responses:
        "200":
          description: Обновлённый список
          content:
            application/json:
              schema: { $ref: "#/components/schemas/List" }
        "400": { description: Некорректное название }
        "403": { description: Вы не владелец }
        "404": { description: Нет такого списка }
    delete:
      summary: Удалить список (только владелец)
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Удалён }
        "403": { description: Вы не владелец }
        "404": { description: Нет такого списка }
  /api/v1/lists/{listID}/members:
    parameters:
        - { name: listID, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Участники списка
      description: Если имя не задано, вместо него приходит замаскированный email (o***@gmail.com).
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Участники
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Member" }
        "404": { description: Нет такого списка }
  /api/v1/lists/{listID}/members/{userID}:
    parameters:
        - { name: listID, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: userID, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      summary: Убрать участника (владелец) или выйти из списка (сам участник)
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Готово }
        "403": { description: Убирать других может только владелец }
        "404": { description: Нет такого списка или участника }
        "409": { description: Владелец не может выйти из списка, его можно только удалить }
  /api/v1/lists/{listID}/invites:
    parameters:
        - { name: listID, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Создать приглашение (может любой участник)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ttl_hours: { type: integer, description: "Срок действия, по умолчанию 168 (7 дней), максимум 720" }
                max_uses: { type: integer, description: "Сколько человек могут войти, по умолчанию 10, максимум 50" }
      responses:
        "201":
          description: Код приглашения
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invite" }
        "404": { description: Нет такого списка }
  /api/v1/invites/accept:
    post:
      summary: Вступить в список по коду приглашения
      description: Код не зависит от регистра. Повторное принятие участником безопасно и не тратит приглашение. В списке не больше 50 участников.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, example: K7M2QX9P }
      responses:
        "200":
          description: Список, в который вы вступили
          content:
            application/json:
              schema: { $ref: "#/components/schemas/List" }
        "404": { description: Код неверный, просрочен, исчерпан или список удалён }
        "409": { description: В списке уже максимум участников }
  /api/v1/lists/{listID}/items:
    parameters:
        - { name: listID, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Позиции списка
      description: >
        Без since возвращает текущие (не удалённые) позиции. С since=N возвращает всё, что изменилось после версии N,
        включая удалённые (deleted=true), чтобы клиент мог убрать их у себя. Значение cursor из ответа передаётся
        как since в следующий запрос. Версии растут строго по порядку внутри списка.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: since, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        "200":
          description: Позиции и курсор
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Item" }
                  cursor: { type: integer, format: int64 }
        "400": { description: Некорректный since }
        "404": { description: Нет такого списка или вы в нём не участвуете }
  /api/v1/lists/{listID}/items/{itemID}:
    parameters:
        - { name: listID, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: itemID, in: path, required: true, schema: { type: string, format: uuid }, description: Id создаёт клиент }
    put:
      summary: Создать позицию или заменить её целиком (идемпотентно)
      description: >
        Подходит для офлайн-клиента: id генерируется на устройстве, повтор запроса безопасен.
        Поля, которых нет в теле, очищаются. Может любой участник списка.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ItemInput" }
      responses:
        "200":
          description: Позиция
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Item" }
        "400": { description: Некорректные данные (название, количество, цена, неизвестная единица или категория) }
        "404": { description: Нет такого списка }
        "409": { description: Позиция с таким id уже есть в другом списке или была удалена }
    patch:
      summary: Изменить часть полей позиции
      description: >
        Меняются только поля, присутствующие в теле. null очищает quantity, unit, price и category_id.
        is_bought=true запоминает, кто купил (повторная отметка другим человеком это не меняет), false очищает.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                quantity: { type: number, nullable: true }
                unit: { type: string, nullable: true }
                price: { type: number, nullable: true }
                category_id: { type: integer, nullable: true }
                position: { type: integer }
                is_bought: { type: boolean }
      responses:
        "200":
          description: Обновлённая позиция
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Item" }
        "400": { description: Некорректные данные }
        "404": { description: Нет такого списка или позиции }
    delete:
      summary: Удалить позицию
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Удалена }
        "404": { description: Нет такого списка или позиции }
  /api/v1/lists/{listID}/items/clear-bought:
    parameters:
        - { name: listID, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Очистить купленное
      description: Купленные позиции убираются из списка и сохраняются в истории покупок. Другие клиенты увидят их как deleted.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Сколько позиций убрано
          content:
            application/json:
              schema:
                type: object
                properties:
                  cleared: { type: integer }
        "404": { description: Нет такого списка }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ItemInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        quantity: { type: number, nullable: true, description: "Больше 0, до 9999999" }
        unit: { type: string, nullable: true, enum: [pcs, kg, g, l, ml, pack] }
        price: { type: number, nullable: true, minimum: 0 }
        category_id: { type: integer, nullable: true }
        catalog_item_id: { type: integer, nullable: true, description: Ссылка на товар справочника }
        position: { type: integer, minimum: 0, default: 0 }
        is_bought: { type: boolean, default: false }
    Item:
      type: object
      properties:
        id: { type: string, format: uuid }
        list_id: { type: string, format: uuid }
        name: { type: string }
        quantity: { type: number, nullable: true }
        unit: { type: string, nullable: true }
        price: { type: number, nullable: true }
        category_id: { type: integer, nullable: true }
        catalog_item_id: { type: integer, nullable: true }
        is_bought: { type: boolean }
        bought_by: { type: string, format: uuid, nullable: true }
        bought_at: { type: string, format: date-time, nullable: true }
        position: { type: integer }
        version: { type: integer, format: int64, description: Растёт при каждом изменении внутри списка }
        updated_at: { type: string, format: date-time }
        deleted: { type: boolean, description: true у удалённых позиций в ответе на since }
    List:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        owner_id: { type: string, format: uuid }
        role: { type: string, enum: [owner, editor], description: Ваша роль в списке }
        member_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Member:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        name: { type: string }
        role: { type: string, enum: [owner, editor] }
        joined_at: { type: string, format: date-time }
    Invite:
      type: object
      properties:
        code: { type: string, example: K7M2QX9P }
        expires_at: { type: string, format: date-time }
        max_uses: { type: integer }
        uses: { type: integer }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
    Tokens:
      type: object
      properties:
        access_token: { type: string, description: JWT, живёт 15 минут }
        refresh_token: { type: string, description: Одноразовый, живёт 90 дней }
        expires_in: { type: integer, example: 900 }
        user: { $ref: "#/components/schemas/User" }
    CatalogItem:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        category_id: { type: integer, nullable: true }
        category_name: { type: string, nullable: true }
        default_unit: { type: string, nullable: true }
    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        sort_order: { type: integer }